major directory restructure
