+++ models/factories/index.js
@@ -1,0 +1,2 @@
+export { PropertyFactory } from './PropertyFactory.js'
+export { ArrayFactory } from './ArrayFactory.js'
