+++ models/index.js
@@ -1,17 +1,1 @@
 export { Gvas } from './Gvas.js'
-export { Property } from './Property.js'
-export { BoolProperty } from './BoolProperty.js'
-export { IntProperty } from './IntProperty.js'
-export { FloatProperty } from './FloatProperty.js'
-export { StrProperty } from './StrProperty.js'
-export { ObjectProperty } from './ObjectProperty.js'
-export { SoftObjectProperty } from './SoftObjectProperty.js'
-export { StructProperty } from './StructProperty.js';
-export { ArrayProperty } from './ArrayProperty.js';
-export { EnumProperty } from './EnumProperty.js';
-export { TupleProperty } from './TupleProperty.js';
-export { Guid } from './Guid.js'
-export { IntArray } from './IntArray.js'
-export { SoftObjectArray } from './SoftObjectArray.js'
-export { StructArray } from './StructArray.js'
-export { PropertyFactory } from './PropertyFactory.js'
