+++ models/properties/BoolProperty.js
