+++ models/properties/FloatProperty.js
