+++ models/properties/IntProperty.js
