+++ models/properties/SoftObjectProperty.js
