+++ models/properties/StrProperty.js
