+++ models/properties/StructArray.js
